requirements.txt
memory-profiler==0.61.0
torchtext==0.12.0
scikit-learn==1.1.3
matplotlib==3.6.2
h5py==3.7.0
tqdm==4.64.1
pytorch-pretrained-bert==0.6.1
requests==2.18.4
urllib3==1.22
pathlib==1.0.1
boto3==1.15.18
botocore==1.18.18
termcolor==2.1.1